+++ src/Http/Resources/BankTransferResource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ChrisLorando\LaravelAccurate\Http\Resources;
+
+class BankTransferResource extends Resource
+{
+    protected string $resourceName = 'bank-transfer';
+}