+++ src/Http/Resources/OtherDepositResource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ChrisLorando\LaravelAccurate\Http\Resources;
+
+class OtherDepositResource extends Resource
+{
+    protected string $resourceName = 'other-deposit';
+}