+++ src/Http/Resources/OtherPaymentResource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ChrisLorando\LaravelAccurate\Http\Resources;
+
+class OtherPaymentResource extends Resource
+{
+    protected string $resourceName = 'other-payment';
+}